+++ ASCOfficePPTFile/PPTFormatLib/Reader/PPTDocumentInfoOneUser.cpp
@@ -198,7 +198,7 @@ bool CPPTUserInfo::ReadFromStream(CRecordUserEditAtom* pUser, POLE::Stream* pStr
 
 void CPPTUserInfo::DecryptStream(POLE::Stream *pStream, int block)
 {
-	int size = pStream->size() - pStream->tell();
+	long size = long(pStream->size() - pStream->tell());
 	
 	POLE::Stream *pStreamTmp = new POLE::Stream(m_pStorageDecrypt, L"Tmp" + std::to_wstring(m_arStreamDecrypt.size() + 1), true, size);
 	unsigned char* data_stream = new unsigned char[size];
