+++ ASCOfficeXlsFile2/source/XlsXlsxConverter/ConvertShapes/CustomShape.h
@@ -105,7 +105,7 @@ public:
 
 	virtual bool SetAdjustment(long index, long value)
 	{
-		if (index < m_arAdjustments.size() && index >= 0)
+		if (index < (long)m_arAdjustments.size() && index >= 0)
 		{
 			m_arAdjustments[index] =  value;
 			return TRUE;
